Change Add to Cart to Set Quantity in cart and add hold/comp to cart (#507)
* WIP

* WIP

* Set quantity on cart rather than add/remove

* Updated cart tests

* WIP

* Fix unit tests

* Update tests

* Add custom CORS header

* Slight readablity fix

* - Added ability to clear event fee
- Removed fee_schedule_id from OrgEdittableAttributes
- Removed the ability to update a fee schedule. The user must use the `create` method to replace the current one

* Add check constraint for order items

* removed dead code

* Find the ticket in the order instead of just taking the first one

* Fixed test

* Order items in the cart

* Added status check

* Remove unused import

* Fix status tests

+++ db/src/test/builders/event_builder.rs
@@ -88,8 +88,8 @@ impl<'a> EventBuilder<'a> {
     }
 
     pub fn with_event_fee(mut self) -> Self {
-        let x: i64 = random();
-        self.fee_in_cents = Some(x % 500);
+        let x: u32 = random();
+        self.fee_in_cents = Some(x as i64 % 500);
         self
     }
